refactor: extract sandbox specification
